requirements.txt
requests
PyYAML
PyGithub
# sqlite3 builtin (no pip package)
python-dotenv